Fix up for map/log verifier split
