+++ merkle/logverifier/hash_chainer.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package verifier
+package logverifier
 
 import (
 	"github.com/google/trillian/merkle/hashers"
